fix(protocol): reject cyclic values and organize sources
