+++ packages/protocol/src/cbor/index.ts
@@ -1,9 +1,9 @@
-export { decodeCbor } from "./cbor-decoder.ts";
-export { encodeCbor } from "./cbor-encoder.ts";
+export { decodeCbor } from "./decoder.ts";
+export { encodeCbor } from "./encoder.ts";
 export {
 	CborError,
 	type CborOptions,
 	DEFAULT_MAX_CBOR_BYTE_LENGTH,
 	DEFAULT_MAX_CBOR_CONTAINER_LENGTH,
 	DEFAULT_MAX_CBOR_DEPTH,
-} from "./cbor-options.ts";
+} from "./types.ts";
