+++ packages/protocol/src/codec.ts
@@ -1,5 +1,5 @@
 import { Check } from "typebox/value";
-import { decodeCbor, encodeCbor } from "./cbor.ts";
+import { decodeCbor, encodeCbor } from "./cbor/index.ts";
 import {
 	assertCompleteFrame,
 	DEFAULT_MAX_FRAME_LENGTH,
@@ -7,8 +7,13 @@ import {
 	FrameDecoder,
 	type FrameDecoderOptions,
 } from "./framing.ts";
-import { type ClientMessage, ClientMessageSchema, type ServerMessage, ServerMessageSchema } from "./messages.ts";
-import { PROTOCOL_VERSION } from "./schemas.ts";
+import {
+	type ClientMessage,
+	ClientMessageSchema,
+	PROTOCOL_VERSION,
+	type ServerMessage,
+	ServerMessageSchema,
+} from "./schemas.ts";
 
 export class ProtocolValidationError extends Error {
 	constructor(message: string, _value?: unknown) {
@@ -17,25 +22,31 @@ export class ProtocolValidationError extends Error {
 	}
 }
 
-function isProtocolValue(value: unknown, optionalProperty = false): boolean {
+function isProtocolValue(value: unknown, optionalProperty = false, ancestors = new Set<object>()): boolean {
 	if (value === undefined) return optionalProperty;
 	if (value === null || typeof value === "boolean" || typeof value === "number" || typeof value === "string") {
 		return true;
 	}
-	if (Array.isArray(value)) return value.every((item) => isProtocolValue(item));
-	if (typeof value !== "object" || Object.getPrototypeOf(value) !== Object.prototype) return false;
-	return Object.values(value).every((item) => isProtocolValue(item, true));
+	if (typeof value !== "object" || ancestors.has(value)) return false;
+	ancestors.add(value);
+	try {
+		if (Array.isArray(value)) return value.every((item) => isProtocolValue(item, false, ancestors));
+		if (Object.getPrototypeOf(value) !== Object.prototype) return false;
+		return Object.values(value).every((item) => isProtocolValue(item, true, ancestors));
+	} finally {
+		ancestors.delete(value);
+	}
 }
 
 export function parseClientMessage(value: unknown): ClientMessage {
-	if (!Check(ClientMessageSchema, value) || !isProtocolValue(value)) {
+	if (!isProtocolValue(value) || !Check(ClientMessageSchema, value)) {
 		throw new ProtocolValidationError("Invalid client protocol message");
 	}
 	return value;
 }
 
 export function parseServerMessage(value: unknown): ServerMessage {
-	if (!Check(ServerMessageSchema, value) || !isProtocolValue(value)) {
+	if (!isProtocolValue(value) || !Check(ServerMessageSchema, value)) {
 		throw new ProtocolValidationError("Invalid server protocol message");
 	}
 	return value;
