+++ packages/protocol/src/index.ts
@@ -1,5 +1,4 @@
-export * from "./cbor.ts";
+export * from "./cbor/index.ts";
 export * from "./codec.ts";
 export * from "./framing.ts";
-export * from "./messages.ts";
 export * from "./schemas.ts";
