+++ packages/protocol/src/schemas.ts
@@ -243,3 +243,170 @@ export const ProtocolErrorSchema = StrictObject({
 });
 export type ProtocolErrorCode = Static<typeof ProtocolErrorCodeSchema>;
 export type ProtocolError = Static<typeof ProtocolErrorSchema>;
+
+const PromptPayloadProperties = {
+	sessionId: IdSchema,
+	text: Type.String(),
+} as const;
+
+export const ListCommandSchema = StrictObject({ command: Type.Literal("list") });
+export const CreateCommandSchema = StrictObject({
+	command: Type.Literal("create"),
+	cwd: Type.Optional(Type.String({ minLength: 1 })),
+	name: Type.Optional(Type.String()),
+	model: Type.Optional(ModelRefSchema),
+	thinkingLevel: Type.Optional(ThinkingLevelSchema),
+});
+export const AttachCommandSchema = StrictObject({ command: Type.Literal("attach"), sessionId: IdSchema });
+export const DetachCommandSchema = StrictObject({ command: Type.Literal("detach"), sessionId: IdSchema });
+export const PromptCommandSchema = StrictObject({ command: Type.Literal("prompt"), ...PromptPayloadProperties });
+export const SteerCommandSchema = StrictObject({ command: Type.Literal("steer"), ...PromptPayloadProperties });
+export const AbortCommandSchema = StrictObject({ command: Type.Literal("abort"), sessionId: IdSchema });
+export const SetModelCommandSchema = StrictObject({
+	command: Type.Literal("set_model"),
+	sessionId: IdSchema,
+	model: ModelRefSchema,
+});
+export const SetThinkingCommandSchema = StrictObject({
+	command: Type.Literal("set_thinking"),
+	sessionId: IdSchema,
+	thinkingLevel: ThinkingLevelSchema,
+});
+export const CommandSchema = Type.Union([
+	ListCommandSchema,
+	CreateCommandSchema,
+	AttachCommandSchema,
+	DetachCommandSchema,
+	PromptCommandSchema,
+	SteerCommandSchema,
+	AbortCommandSchema,
+	SetModelCommandSchema,
+	SetThinkingCommandSchema,
+]);
+export type Command = Static<typeof CommandSchema>;
+export type CommandName = Command["command"];
+
+export const CreateResultSchema = StrictObject({
+	command: Type.Literal("create"),
+	session: SessionSnapshotSchema,
+});
+export const AttachResultSchema = StrictObject({
+	command: Type.Literal("attach"),
+	session: SessionSnapshotSchema,
+});
+export const PromptResultSchema = StrictObject({
+	command: Type.Literal("prompt"),
+	session: SessionSnapshotSchema,
+});
+export const SteerResultSchema = StrictObject({
+	command: Type.Literal("steer"),
+	session: SessionSnapshotSchema,
+});
+export const AbortResultSchema = StrictObject({
+	command: Type.Literal("abort"),
+	session: SessionSnapshotSchema,
+});
+export const SetModelResultSchema = StrictObject({
+	command: Type.Literal("set_model"),
+	session: SessionSnapshotSchema,
+});
+export const SetThinkingResultSchema = StrictObject({
+	command: Type.Literal("set_thinking"),
+	session: SessionSnapshotSchema,
+});
+
+export const ListResultSchema = StrictObject({
+	command: Type.Literal("list"),
+	sessions: Type.Array(SessionSummarySchema),
+});
+export const DetachResultSchema = StrictObject({
+	command: Type.Literal("detach"),
+	sessionId: IdSchema,
+});
+export const CommandResultSchema = Type.Union([
+	ListResultSchema,
+	CreateResultSchema,
+	AttachResultSchema,
+	DetachResultSchema,
+	PromptResultSchema,
+	SteerResultSchema,
+	AbortResultSchema,
+	SetModelResultSchema,
+	SetThinkingResultSchema,
+]);
+export type CommandResult = Static<typeof CommandResultSchema>;
+
+export type ResultForCommand<TCommand extends Command> = TCommand["command"] extends "list"
+	? Static<typeof ListResultSchema>
+	: TCommand["command"] extends "detach"
+		? Static<typeof DetachResultSchema>
+		: Extract<CommandResult, { command: TCommand["command"] }>;
+
+/** Must be the first frame sent by a client. Version is intentionally an integer, not a coercible string. */
+export const ClientHelloSchema = StrictObject({
+	type: Type.Literal("hello"),
+	version: Type.Integer({ minimum: 0 }),
+	token: Type.String({ minLength: 1 }),
+});
+export type ClientHello = Static<typeof ClientHelloSchema>;
+
+export const RequestEnvelopeSchema = StrictObject({
+	type: Type.Literal("request"),
+	id: IdSchema,
+	request: CommandSchema,
+});
+export type RequestEnvelope = Static<typeof RequestEnvelopeSchema>;
+export const ClientMessageSchema = Type.Union([ClientHelloSchema, RequestEnvelopeSchema]);
+export type ClientMessage = Static<typeof ClientMessageSchema>;
+
+export const ServerEventSchema = Type.Union([
+	StrictObject({ type: Type.Literal("server_snapshot"), snapshot: ServerSnapshotSchema }),
+	StrictObject({ type: Type.Literal("session_snapshot"), snapshot: SessionSnapshotSchema }),
+	StrictObject({
+		type: Type.Literal("session_progress"),
+		sessionId: IdSchema,
+		progress: TranscriptProgressSchema,
+	}),
+	StrictObject({ type: Type.Literal("session_removed"), sessionId: IdSchema }),
+]);
+export type ServerEvent = Static<typeof ServerEventSchema>;
+
+export const ServerHelloSchema = StrictObject({
+	type: Type.Literal("hello"),
+	version: Type.Literal(PROTOCOL_VERSION),
+	connectionId: IdSchema,
+	snapshot: ServerSnapshotSchema,
+});
+export const ServerHelloErrorSchema = StrictObject({
+	type: Type.Literal("hello_error"),
+	error: ProtocolErrorSchema,
+});
+export const ResponseEnvelopeSchema = Type.Union([
+	StrictObject({
+		type: Type.Literal("response"),
+		id: IdSchema,
+		ok: Type.Literal(true),
+		result: CommandResultSchema,
+	}),
+	StrictObject({
+		type: Type.Literal("response"),
+		id: IdSchema,
+		ok: Type.Literal(false),
+		error: ProtocolErrorSchema,
+	}),
+]);
+export const EventEnvelopeSchema = StrictObject({
+	type: Type.Literal("event"),
+	event: ServerEventSchema,
+});
+export const ServerMessageSchema = Type.Union([
+	ServerHelloSchema,
+	ServerHelloErrorSchema,
+	ResponseEnvelopeSchema,
+	EventEnvelopeSchema,
+]);
+export type ServerHello = Static<typeof ServerHelloSchema>;
+export type ServerHelloError = Static<typeof ServerHelloErrorSchema>;
+export type ResponseEnvelope = Static<typeof ResponseEnvelopeSchema>;
+export type EventEnvelope = Static<typeof EventEnvelopeSchema>;
+export type ServerMessage = Static<typeof ServerMessageSchema>;
